+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Support for `SourceConfig.description` property to allow setting a description for the source
+
 ## [0.10.0] (2023-09-04)
 
 ### Added