+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Support to get thumbnail images via `Player.getThumbnail` and `Source.getThumbnail`
+- `Player.getThumbnail` and `Source.getThumbnail` APIs to get thumbnail images
 
 ## [0.10.0] (2023-09-04)